moved shared test stuff to conftest.py
